refactor: use ctx to carry metrics args
