+++ moat/src/metrics.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{borrow::Cow, sync::LazyLock, time::Instant};
+use std::{borrow::Cow, sync::LazyLock};
 
 use opentelemetry::{
     KeyValue,
@@ -114,7 +114,7 @@ impl S3Metrics {
         Self {
             count: meter
                 .u64_counter("moat.s3.count")
-                .with_description("Moat API call count")
+                .with_description("Moat S3 API call count")
                 .build(),
             bytes: meter
                 .u64_counter("moat.s3.bytes")
@@ -122,7 +122,7 @@ impl S3Metrics {
                 .build(),
             duration: meter
                 .f64_histogram("moat.s3.duration")
-                .with_description("Moat API call duration in seconds")
+                .with_description("Moat S3 API call duration in seconds")
                 .with_unit("second")
                 .with_boundaries(vec![
                     0.0001, 0.0005, 0.001, 0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.5, 1.0, 5.0, 10.0,
@@ -134,89 +134,23 @@ impl S3Metrics {
     pub fn labels(
         operation: impl Into<Cow<'static, str>>,
         status: impl Into<Cow<'static, str>>,
-        ctx: impl Into<Cow<'static, str>>,
+        extra: impl Into<Cow<'static, str>>,
     ) -> [KeyValue; 3] {
         [
             KeyValue::new("operation", operation.into()),
             KeyValue::new("status", status.into()),
-            KeyValue::new("ctx", ctx.into()),
+            KeyValue::new("extra", extra.into()),
         ]
     }
 
     pub const OPERATION_GET_OBJECT: &'static str = "GetObject";
 
     pub const STATUS_OK: &'static str = "ok";
+    pub const STATUS_ERR: &'static str = "err";
 
-    pub const CTX_NONE: &'static str = "none";
-    pub const CTX_CACHED: &'static str = "cached";
-    pub const CTX_FETCHED: &'static str = "fetched";
-    pub const CTX_PROXIED: &'static str = "proxied";
-    pub const CTX_S3: &'static str = "s3";
-}
-
-#[derive(Debug)]
-pub struct S3MetricsGuard {
-    operation: Cow<'static, str>,
-    status: Cow<'static, str>,
-    ctx: Cow<'static, str>,
-    count: u64,
-    bytes: u64,
-    start: Instant,
-}
-
-impl Drop for S3MetricsGuard {
-    fn drop(&mut self) {
-        let metrics = Metrics::global();
-        metrics.s3.count.add(
-            self.count,
-            &S3Metrics::labels(self.operation.clone(), self.status.clone(), self.ctx.clone()),
-        );
-        metrics.s3.bytes.add(
-            self.bytes,
-            &S3Metrics::labels(self.operation.clone(), self.status.clone(), self.ctx.clone()),
-        );
-        metrics.s3.duration.record(
-            self.start.elapsed().as_secs_f64(),
-            &S3Metrics::labels(self.operation.clone(), self.status.clone(), self.ctx.clone()),
-        );
-    }
-}
-
-impl S3MetricsGuard {
-    pub fn new(
-        operation: impl Into<Cow<'static, str>>,
-        status: impl Into<Cow<'static, str>>,
-        ctx: impl Into<Cow<'static, str>>,
-        count: u64,
-        bytes: u64,
-    ) -> Self {
-        let operation = operation.into();
-        let status = status.into();
-        let ctx = ctx.into();
-        let start = Instant::now();
-        S3MetricsGuard {
-            operation,
-            status,
-            ctx,
-            count,
-            bytes,
-            start,
-        }
-    }
-
-    #[expect(unused)]
-    pub fn status(&mut self, status: impl Into<Cow<'static, str>>) -> &mut Self {
-        self.status = status.into();
-        self
-    }
-
-    pub fn ctx(&mut self, ctx: impl Into<Cow<'static, str>>) -> &mut Self {
-        self.ctx = ctx.into();
-        self
-    }
-
-    pub fn add_bytes(&mut self, bytes: u64) -> &mut Self {
-        self.bytes += bytes;
-        self
-    }
+    pub const EXTRA_NONE: &'static str = "none";
+    pub const EXTRA_CACHED: &'static str = "cached";
+    pub const EXTRA_FETCHED: &'static str = "fetched";
+    pub const EXTRA_PROXIED: &'static str = "proxied";
+    pub const EXTRA_S3: &'static str = "s3";
 }
